+++ arch/x86/include/kernel_arch_func.h
@@ -47,6 +47,27 @@ void z_x86_early_serial_init(void);
 void z_x86_paging_init(void);
 #endif /* CONFIG_X86_MMU */
 
-#endif
+/* Called upon unrecoverable error; dump registers and transfer control to
+ * kernel via z_fatal_error()
+ */
+FUNC_NORETURN void z_x86_fatal_error(unsigned int reason,
+				     const z_arch_esf_t *esf);
+
+#ifdef CONFIG_THREAD_STACK_INFO
+/**
+ * @brief Check if a memory address range falls within the stack
+ *
+ * Given a memory address range, ensure that it falls within the bounds
+ * of the faulting context's stack.
+ *
+ * @param addr Starting address
+ * @param size Size of the region, or 0 if we just want to see if addr is
+ *             in bounds
+ * @param cs Code segment of faulting context
+ * @return true if addr/size region is not within the thread stack
+ */
+bool z_x86_check_stack_bounds(uintptr_t addr, size_t size, u16_t cs);
+#endif /* CONFIG_THREAD_STACK_INFO */
+#endif /* !_ASMLANGUAGE */
 
 #endif /* ZEPHYR_ARCH_X86_INCLUDE_KERNEL_ARCH_FUNC_H_ */
